pyproject.toml
[tool.poetry]
name = "memecry"
version = "0.1.0"
description = ""
authors = ["crpier <crpier42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.0b1"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
